Moved testing buttons
